refactor: moved feature-specific code from shared
