+++ src/index.ts
@@ -3,6 +3,7 @@ import type { ESLint, Linter } from "eslint"
 import { name, version } from "../package.json"
 
 import enforceEffectNamingConvention from "./rules/enforce-effect-naming-convention/enforce-effect-naming-convention"
+import enforceExhaustiveUseUnitDestructuring from "./rules/enforce-exhaustive-useUnit-destructuring/enforce-exhaustive-useUnit-destructuring"
 import enforceGateNamingConvention from "./rules/enforce-gate-naming-convention/enforce-gate-naming-convention"
 import enforceStoreNamingConvention from "./rules/enforce-store-naming-convention/enforce-store-naming-convention"
 import keepOptionsOrder from "./rules/keep-options-order/keep-options-order"
@@ -21,7 +22,6 @@ import noUnnecessaryDuplication from "./rules/no-unnecessary-duplication/no-unne
 import noUselessMethods from "./rules/no-useless-methods/no-useless-methods"
 import noWatch from "./rules/no-watch/no-watch"
 import preferUseUnit from "./rules/prefer-useUnit/prefer-useUnit"
-import preferUseUnitDestructuring from "./rules/prefer-useUnit-destructuring/prefer-useUnit-destructuring"
 import requirePickupInPersist from "./rules/require-pickup-in-persist/require-pickup-in-persist"
 import strictEffectHandlers from "./rules/strict-effect-handlers/strict-effect-handlers"
 import { ruleset } from "./ruleset"
@@ -30,6 +30,7 @@ const base = {
   meta: { name, version, namespace: "effector" },
   rules: {
     "enforce-effect-naming-convention": enforceEffectNamingConvention,
+    "enforce-exhaustive-useUnit-destructuring": enforceExhaustiveUseUnitDestructuring,
     "enforce-gate-naming-convention": enforceGateNamingConvention,
     "enforce-store-naming-convention": enforceStoreNamingConvention,
     "keep-options-order": keepOptionsOrder,
@@ -48,7 +49,6 @@ const base = {
     "no-useless-methods": noUselessMethods,
     "no-watch": noWatch,
     "prefer-useUnit": preferUseUnit,
-    "prefer-useUnit-destructuring": preferUseUnitDestructuring,
     "require-pickup-in-persist": requirePickupInPersist,
     "strict-effect-handlers": strictEffectHandlers,
   },
